wip: migrating to rollup all the methods in each damn file
