+++ lib/browser/tag/styleManager.js
@@ -1,11 +1,14 @@
+import { $, mkEl, setAttr } from '../util'
+
 /**
  * Object that will be used to inject and manage the css of every tag instance
  */
-export default (function(_riot) {
+export default (function() {
 
   if (!window) return { // skip injection on the server
     add: function () {},
-    inject: function () {}
+    inject: function () {},
+    styleNode: null
   }
 
   var styleNode = (function () {
@@ -28,16 +31,11 @@ export default (function(_riot) {
   var cssTextProp = styleNode.styleSheet,
     stylesToInject = ''
 
-  // Expose the style node in a non-modificable property
-  Object.defineProperty(_riot, 'styleNode', {
-    value: styleNode,
-    writable: true
-  })
-
   /**
    * Public api
    */
   return {
+    styleNode: styleNode,
     /**
      * Save a tag style to be later injected into DOM
      * @param   { String } css [description]
@@ -58,4 +56,4 @@ export default (function(_riot) {
     }
   }
 
-})(riot)
+}())
