+++ lib/browser/tag/tag.js
@@ -1,6 +1,7 @@
 import observable from 'riot-observable'
 import parseExpressions from './parse'
 import { update } from './update'
+import { inherit, cleanUpData, defineProperty } from '../util'
 
 
 export default function Tag(impl, conf, innerHTML) {
