+++ lib/browser/util.js
@@ -1,3 +1,5 @@
+import { __tagImpl, T_OBJECT } from './global-variables'
+import Tag from './tag/tag'
 /**
  * Specialized function for looping an array-like collection with `each={}`
  * @param   { Array } els - collection of items
