+++ test/tags.html
@@ -78,7 +78,7 @@
 
   </div>
 
-  <script src="../dist/riot/riot.js"></script>
+  <script src="../dist/riot/riot.rollup.js"></script>
   <script src="../dist/tags.js"></script>
   <script>var tags = riot.mount('*', { test: { zap: 100 }})</script>
